dom/workers/test/test_blobWorkers.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
  <head>
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js">
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <script type="text/javascript">
      const message = "hi";

      const workerScript =
        "onmessage = function(event) {" +
        "  postMessage(event.data);" +
        "};";

      var bb = new MozBlobBuilder();
      bb.append(workerScript);

      var worker = new Worker(URL.createObjectURL(bb.getBlob()));
      worker.onmessage = function(event) {
        is(event.data, message, "Got correct message");
        SimpleTest.finish();
      };
      worker.postMessage(message);

      SimpleTest.waitForExplicitFinish();
    </script>
  </body>
</html>
